add tests and clean up

+++ src/ProductsApiClient/Models/Product.cs
@@ -4,4 +4,4 @@ public record Product(
     Guid Id,
     string Name,
     string Supplier,
-    double UnitPrice);
+    decimal UnitPrice);
